+++ src/main/java/edu/ie3/datamodel/models/input/system/type/BmTypeInput.java
@@ -114,12 +114,12 @@ public class BmTypeInput extends SystemParticipantTypeInput {
     public BmTypeInputCopyBuilder setActivePowerGradient(
         ComparableQuantity<DimensionlessRate> activePowerGradient) {
       this.activePowerGradient = activePowerGradient;
-      return this;
+      return thisInstance();
     }
 
     public BmTypeInputCopyBuilder setEtaConv(ComparableQuantity<Dimensionless> etaConv) {
       this.etaConv = etaConv;
-      return this;
+      return thisInstance();
     }
 
     public ComparableQuantity<DimensionlessRate> getActivePowerGradient() {
@@ -134,7 +134,7 @@ public class BmTypeInput extends SystemParticipantTypeInput {
     public BmTypeInputCopyBuilder scale(Double factor) {
       setCapex(getCapex().multiply(factor));
       setsRated(getsRated().multiply(factor));
-      return this;
+      return thisInstance();
     }
 
     @Override